@@ -1,8 +1,9 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-	<TargetFramework>net6.0</TargetFramework>
+	<TargetFramework>net8.0</TargetFramework>
 	<ImplicitUsings>enable</ImplicitUsings>
+	<Nullable>enable</Nullable>
 
     <PackageId>Cofoundry.Plugins.Mail.SendGrid</PackageId>
     <Description>Cofoundry mail service implemention for SendGrid.</Description>
@@ -12,13 +13,13 @@
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-    <DocumentationFile>bin\Release\net6.0\Cofoundry.Plugins.Mail.SendGrid.xml</DocumentationFile>
+    <DocumentationFile>bin\Release\net8.0\Cofoundry.Plugins.Mail.SendGrid.xml</DocumentationFile>
     <NoWarn>1701;1702;1705;1591</NoWarn>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Cofoundry.Core" Version="0.11.1" />
-    <PackageReference Include="Sendgrid" Version="9.26.0" />
+    <PackageReference Include="Cofoundry.Core" Version="0.11.5-dotnet-8-0031" />
+    <PackageReference Include="Sendgrid" Version="9.29.3" />
   </ItemGroup>
 
 </Project>
