@@ -1,4 +1,4 @@
-﻿using Cofoundry.Core.DependencyInjection;
+using Cofoundry.Core.DependencyInjection;
 using Cofoundry.Core.Mail;
 using Cofoundry.Plugins.Mail.SendGrid.Internal;
 
@@ -8,7 +8,10 @@ public class SendGridDependencyRegistration : IDependencyRegistration
 {
     public void Register(IContainerRegister container)
     {
-        if (container.Configuration.GetValue<bool>("Cofoundry:Plugins:SendGrid:Disabled")) return;
+        if (container.Configuration.GetValue<bool>("Cofoundry:Plugins:SendGrid:Disabled"))
+        {
+            return;
+        }
 
         var overrideOptions = RegistrationOptions.Override();
 
