@@ -1,5 +1,5 @@
-﻿using Cofoundry.Core.Configuration;
 using System.ComponentModel.DataAnnotations;
+using Cofoundry.Core.Configuration;
 
 namespace Cofoundry.Plugins.Mail.SendGrid;
 
@@ -7,7 +7,7 @@ public class SendGridSettings : PluginConfigurationSettingsBase
 {
     /// <summary>
     /// Indicates whether the plugin should be disabled, which means services
-    /// will not be bootstrapped. Defaults to false.
+    /// will not be bootstrapped. Defaults to <see langword="false"/>.
     /// </summary>
     public bool Disabled { get; set; }
 
@@ -15,5 +15,5 @@ public class SendGridSettings : PluginConfigurationSettingsBase
     /// The api key use when authenticating with the SendGrid api.
     /// </summary>
     [Required]
-    public string ApiKey { get; set; }
+    public string ApiKey { get; set; } = string.Empty;
 }
